Use json.dumps() for settings set by nikola init.
Previously, all variables were strings that were substituted
into X = ${y} or X = "${y}", depending on the intended
type of the variable.  As of now, all but three variables (which
contain comments and are formatted specially) are handled by
json.dumps(), which outputs properly escaped code that is substituted
into expressions of type 2 above.

Signed-off-by: Chris “Kwpolska” Warrick <kwpolska@gmail.com>